+++ doc/conf.py
@@ -22,7 +22,7 @@ copyright = "2020, Gemeente Utrecht"
 author = "Gemeente Utrecht"
 
 # The full version, including alpha/beta/rc tags
-release = "0.1.0"
+release = "0.1.1"
 
 
 # -- General configuration ---------------------------------------------------