+++ CHANGELOG.rst
@@ -2,10 +2,14 @@
 Change history
 ==============
 
+0.1.1
+=====
+
+* Added the unversioned URL of the document in the DRC, and use this explicitly in several functions related to updating, locking and unlocking of documents on the DRC.
+* Added bulk force_deletion in admin page of Document files.
+* Updated API documentation to reflect all relevant query parameters for the list endpoint.
 
 0.1.0
 =====
-
-*<month, day year>*
 
 * Initial release.